Source code of the PH7 Interpreter
The PH7 interpreter is a simple,basic stand-alone PHP interpreter that allows
the user to enter and execute PHP files against a PH7 engine. 
To start the ph7 program, just type "ph7" followed by the name of the PHP file
to compile and execute. That is, the first argument is to the interpreter, the rest
are scripts arguments, press "Enter" and the PHP code will be executed.
If something goes wrong while processing the PHP script due to a compile-time
error,your error output (STDOUT) should display the compile-time error messages.